+++ src/main/java/com/liviaaurich/locadora/service/impl/ClasseServicoImpl.java
@@ -2,6 +2,7 @@ package com.liviaaurich.locadora.service.impl;
 
 import com.liviaaurich.locadora.domain.Classe;
 import com.liviaaurich.locadora.repository.ClasseRepository;
+import com.liviaaurich.locadora.repository.TituloRepository;
 import com.liviaaurich.locadora.service.BaseService;
 import com.liviaaurich.locadora.service.dto.ClasseDTO;
 import com.liviaaurich.locadora.service.dto.dropdown.DropdownDTO;
@@ -25,10 +26,12 @@ import static org.hibernate.id.IdentifierGenerator.ENTITY_NAME;
 public class ClasseServicoImpl implements BaseService<ClasseDTO> {
 
     private static final String MSG_CLASSE_INEXISTENTE = "Não foi possível obter a Classe. ID não está presente.";
+    private static final String CLASSE = "Classe";
 
     private final ClasseRepository classeRepository;
     private final ClasseMapper classeMapper;
     private final ClasseDropdownMapper classeDropdownMapper;
+    private final TituloRepository tituloRepository;
 
     @Override
     public ClasseDTO salvar(ClasseDTO classeDTO) {
@@ -43,6 +46,10 @@ public class ClasseServicoImpl implements BaseService<ClasseDTO> {
     public void excluir(Long id) {
         Classe classe = classeRepository.findById(id).orElseThrow(() ->
             new BadRequestAlertException(MSG_CLASSE_INEXISTENTE, ENTITY_NAME, "id"));
+
+        if(!tituloRepository.findAllByDiretorId(classe.getId()).isEmpty()) {
+            throw new BadRequestAlertException("A Classe selecionada está vinculada a um Título.", ENTITY_NAME, CLASSE);
+        }
 
         classeRepository.delete(classe);
     }
